Change uvi factor type to float (#27)
* Fix conversion: add float() call

Sometimes the UV data comes as a float string ('0.8'), and it causes `ValueError: invalid literal for int() with base 10: '0.8'`

* Change type of uvi factor to float

+++ weatheril/weather.py
@@ -24,7 +24,7 @@ class Weather:
     u_v_index: int
     u_v_level: str
     u_v_i_max: int
-    u_v_i_factor: int
+    u_v_i_factor: float
     json: str
     weather_code: str
     wave_height: float
